+++ libadwaita_core/lib/libadwaita_core.dart
@@ -2,6 +2,8 @@ library libadwaita_core;
 
 import 'package:flutter/material.dart';
 
+typedef AdwControlsWidget = Widget? Function(VoidCallback?)?;
+
 class AdwControls {
   AdwControls({
     this.closeBtn,
@@ -9,9 +11,20 @@ class AdwControls {
     this.minimizeBtn,
   });
 
-  final Widget? Function(VoidCallback? onClose)? closeBtn;
-  final Widget? Function(VoidCallback? onMaximize)? maximizeBtn;
-  final Widget? Function(VoidCallback? onMinimize)? minimizeBtn;
+  final AdwControlsWidget closeBtn;
+  final AdwControlsWidget maximizeBtn;
+  final AdwControlsWidget minimizeBtn;
+
+  AdwControls copyWith(
+    AdwControlsWidget closeBtn,
+    AdwControlsWidget maximizeBtn,
+    AdwControlsWidget minimizeBtn,
+  ) =>
+      AdwControls(
+        closeBtn: closeBtn ?? this.closeBtn,
+        maximizeBtn: maximizeBtn ?? this.maximizeBtn,
+        minimizeBtn: minimizeBtn ?? this.minimizeBtn,
+      );
 }
 
 class AdwActions {
@@ -21,6 +34,7 @@ class AdwActions {
     this.onMinimize,
     this.onDoubleTap,
     this.onHeaderDrag,
+    this.onRightClick,
   });
 
   final VoidCallback? onClose;
@@ -28,4 +42,22 @@ class AdwActions {
   final VoidCallback? onMinimize;
   final VoidCallback? onDoubleTap;
   final VoidCallback? onHeaderDrag;
+  final VoidCallback? onRightClick;
+
+  AdwActions copyWith(
+    VoidCallback? onClose,
+    VoidCallback? onMaximize,
+    VoidCallback? onMinimize,
+    VoidCallback? onDoubleTap,
+    VoidCallback? onHeaderDrag,
+    VoidCallback? onRightClick,
+  ) =>
+      AdwActions(
+        onClose: onClose ?? this.onClose,
+        onMaximize: onMaximize ?? this.onMaximize,
+        onMinimize: onMinimize ?? this.onMinimize,
+        onDoubleTap: onDoubleTap ?? this.onDoubleTap,
+        onHeaderDrag: onHeaderDrag ?? this.onHeaderDrag,
+        onRightClick: onRightClick ?? this.onRightClick,
+      );
 }