+++ libadwaita_core/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.5.2
 
+* Add copyWith method
+* Add onRightClick property to `AdwActions`
+
+## 0.5.1+1
+
 * Update Documentation
  
 ## 0.5.1